requirements.txt
numpy>=1.24,<2.0

pandas>=2.1

scikit-learn>=1.3

librosa>=0.10.1

soundfile>=0.12

audioread>=3.0.1

tensorflow>=2.16.1,<2.18

protobuf==5.28.3

fastapi>=0.111

uvicorn[standard]>=0.30

python-multipart>=0.0.9

jinja2>=3.1

aiofiles>=23.2

itsdangerous>=2.2

python-dotenv>=1.0

pydantic>=2.6

matplotlib>=3.8

seaborn>=0.13

tqdm>=4.66

requests>=2.31

loguru>=0.7

reportlab>=4.0

SQLAlchemy>=2.0

aiosqlite>=0.19

kaggle>=1.6

kagglesdk>=0.17

gunicorn>=21.2